patch: refactor scanner, loader and flasher out of gui + upgrade to electron 25
