+++ .gitignore
@@ -28,6 +28,7 @@ pids
 
 # Generated files
 /generated
+/binaries
 
 # Dependency directory
 # https://docs.npmjs.com/misc/faq#should-i-check-my-node-modules-folder-into-git
