+++ lib/gui/app/modules/image-writer.ts
@@ -17,38 +17,15 @@
 import { Drive as DrivelistDrive } from 'drivelist';
 import * as sdk from 'etcher-sdk';
 import { Dictionary } from 'lodash';
-import * as ipc from 'node-ipc';
-import * as os from 'os';
-import * as path from 'path';
-
-import * as packageJSON from '../../../../package.json';
 import * as errors from '../../../shared/errors';
-import * as permissions from '../../../shared/permissions';
-import { getAppPath } from '../../../shared/utils';
-import { SourceMetadata } from '../components/source-selector/source-selector';
+import { SourceMetadata } from '../../../shared/typings/source-selector';
 import * as flashState from '../models/flash-state';
 import * as selectionState from '../models/selection-state';
 import * as settings from '../models/settings';
 import * as analytics from '../modules/analytics';
 import * as windowProgress from '../os/window-progress';
-
-const THREADS_PER_CPU = 16;
-
-// There might be multiple Etcher instances running at
-// the same time, therefore we must ensure each IPC
-// server/client has a different name.
-const IPC_SERVER_ID = `etcher-server-${process.pid}`;
-const IPC_CLIENT_ID = `etcher-client-${process.pid}`;
-
-ipc.config.id = IPC_SERVER_ID;
-ipc.config.socketRoot = path.join(
-	process.env.XDG_RUNTIME_DIR || os.tmpdir(),
-	path.sep,
-);
-
-// NOTE: Ensure this isn't disabled, as it will cause
-// the stdout maxBuffer size to be exceeded when flashing
-ipc.config.silent = true;
+import { startApiAndSpawnChild } from './api';
+import { terminateScanningServer } from '../app';
 
 /**
  * @summary Handle a flash error and log it to analytics
@@ -80,51 +57,7 @@ function handleErrorLogging(
 	}
 }
 
-function terminateServer() {
-	// Turns out we need to destroy all sockets for
-	// the server to actually close. Otherwise, it
-	// just stops receiving any further connections,
-	// but remains open if there are active ones.
-	// @ts-ignore (no Server.sockets in @types/node-ipc)
-	for (const socket of ipc.server.sockets) {
-		socket.destroy();
-	}
-	ipc.server.stop();
-}
-
-function writerArgv(): string[] {
-	let entryPoint = path.join(getAppPath(), 'generated', 'child-writer.js');
-	// AppImages run over FUSE, so the files inside the mount point
-	// can only be accessed by the user that mounted the AppImage.
-	// This means we can't re-spawn Etcher as root from the same
-	// mount-point, and as a workaround, we re-mount the original
-	// AppImage as root.
-	if (os.platform() === 'linux' && process.env.APPIMAGE && process.env.APPDIR) {
-		entryPoint = entryPoint.replace(process.env.APPDIR, '');
-		return [
-			process.env.APPIMAGE,
-			'-e',
-			`require(\`\${process.env.APPDIR}${entryPoint}\`)`,
-		];
-	} else {
-		return [process.argv[0], entryPoint];
-	}
-}
-
-function writerEnv() {
-	return {
-		IPC_SERVER_ID,
-		IPC_CLIENT_ID,
-		IPC_SOCKET_ROOT: ipc.config.socketRoot,
-		ELECTRON_RUN_AS_NODE: '1',
-		UV_THREADPOOL_SIZE: (os.cpus().length * THREADS_PER_CPU).toString(),
-		// This environment variable prevents the AppImages
-		// desktop integration script from presenting the
-		// "installation" dialog
-		SKIP: '1',
-		...(process.platform === 'win32' ? {} : process.env),
-	};
-}
+let cancelEmitter: (type: string) => void | undefined;
 
 interface FlashResults {
 	skip?: boolean;
@@ -144,22 +77,13 @@ async function performWrite(
 	drives: DrivelistDrive[],
 	onProgress: sdk.multiWrite.OnProgressFunction,
 ): Promise<{ cancelled?: boolean }> {
-	let cancelled = false;
-	let skip = false;
-	ipc.serve();
 	const { autoBlockmapping, decompressFirst } = await settings.getAll();
-	return await new Promise((resolve, reject) => {
-		ipc.server.on('error', (error) => {
-			terminateServer();
-			const errorObject = errors.fromJSON(error);
-			reject(errorObject);
-		});
 
-		ipc.server.on('log', (message) => {
-			console.log(message);
-		});
+	console.log({ image, drives });
 
+	return await new Promise(async (resolve, reject) => {
 		const flashResults: FlashResults = {};
+
 		const analyticsData = {
 			image,
 			drives,
@@ -168,75 +92,51 @@ async function performWrite(
 			flashInstanceUuid: flashState.getFlashUuid(),
 		};
 
-		ipc.server.on('fail', ({ device, error }) => {
+		const onFail = ({ device, error }) => {
+			console.log('fail event');
+			console.log(device);
+			console.log(error);
 			if (device.devicePath) {
 				flashState.addFailedDeviceError({ device, error });
 			}
 			handleErrorLogging(error, analyticsData);
-		});
+			finish();
+		};
 
-		ipc.server.on('done', (event) => {
+		const onDone = (event) => {
+			console.log('done event');
 			event.results.errors = event.results.errors.map(
 				(data: Dictionary<any> & { message: string }) => {
 					return errors.fromJSON(data);
 				},
 			);
 			flashResults.results = event.results;
-		});
+			finish();
+		};
 
-		ipc.server.on('abort', () => {
-			terminateServer();
-			cancelled = true;
-		});
+		const onAbort = () => {
+			console.log('abort event');
+			flashResults.cancelled = true;
+			finish();
+		};
 
-		ipc.server.on('skip', () => {
-			terminateServer();
-			skip = true;
-		});
+		const onSkip = () => {
+			console.log('skip event');
+			flashResults.skip = true;
+			finish();
+		};
 
-		ipc.server.on('state', onProgress);
-
-		ipc.server.on('ready', (_data, socket) => {
-			ipc.server.emit(socket, 'write', {
-				image,
-				destinations: drives,
-				SourceType: image.SourceType.name,
-				autoBlockmapping,
-				decompressFirst,
-			});
-		});
-
-		const argv = writerArgv();
-
-		ipc.server.on('start', async () => {
-			console.log(`Elevating command: ${argv.join(' ')}`);
-			const env = writerEnv();
-			try {
-				const results = await permissions.elevateCommand(argv, {
-					applicationName: packageJSON.displayName,
-					environment: env,
-				});
-				flashResults.cancelled = cancelled || results.cancelled;
-				flashResults.skip = skip;
-			} catch (error: any) {
-				// This happens when the child is killed using SIGKILL
-				const SIGKILL_EXIT_CODE = 137;
-				if (error.code === SIGKILL_EXIT_CODE) {
-					error.code = 'ECHILDDIED';
-				}
-				reject(error);
-			} finally {
-				console.log('Terminating IPC server');
-				terminateServer();
-			}
+		const finish = () => {
 			console.log('Flash results', flashResults);
 
 			// The flash wasn't cancelled and we didn't get a 'done' event
+			// Catch unexepected situation
 			if (
 				!flashResults.cancelled &&
 				!flashResults.skip &&
 				flashResults.results === undefined
 			) {
+				console.log(flashResults);
 				reject(
 					errors.createUserError({
 						title: 'The writer process ended unexpectedly',
@@ -244,15 +144,40 @@ async function performWrite(
 							'Please try again, and contact the Etcher team if the problem persists',
 					}),
 				);
-				return;
 			}
-			resolve(flashResults);
-		});
 
-		// Clear the update lock timer to prevent longer
-		// flashing timing it out, and releasing the lock
-		ipc.server.start();
+			console.log('Terminating IPC server');
+			terminateServer();
+			resolve(flashResults);
+		};
+
+		// Spawn the child process with privileges and wait for the connection to be made
+		const { emit, registerHandler, terminateServer } =
+			await startApiAndSpawnChild({
+				withPrivileges: true,
+			});
+
+		registerHandler('state', onProgress);
+		registerHandler('fail', onFail);
+		registerHandler('done', onDone);
+		registerHandler('abort', onAbort);
+		registerHandler('skip', onSkip);
+
+		cancelEmitter = (cancelStatus: string) => emit(cancelStatus);
+
+		// Now that we know we're connected we can instruct the child process to start the write
+		const paramaters = {
+			image,
+			destinations: drives,
+			SourceType: image.SourceType,
+			autoBlockmapping,
+			decompressFirst,
+		};
+		console.log('params', paramaters);
+		emit('write', paramaters);
 	});
+
+	// The process continue in the event handler
 }
 
 /**
@@ -269,6 +194,7 @@ export async function flash(
 	}
 
 	await flashState.setFlashingFlag();
+
 	flashState.setDevicePaths(
 		drives.map((d) => d.devicePath).filter((p) => p != null) as string[],
 	);
@@ -284,6 +210,7 @@ export async function flash(
 
 	analytics.logEvent('Flash', analyticsData);
 
+	// start api and call the flasher
 	try {
 		const result = await write(image, drives, flashState.setProgressState);
 		await flashState.unsetFlashingFlag(result);
@@ -292,8 +219,11 @@ export async function flash(
 			cancelled: false,
 			errorCode: error.code,
 		});
+
 		windowProgress.clear();
+
 		const { results = {} } = flashState.getFlashResults();
+
 		const eventData = {
 			...analyticsData,
 			errors: results.errors,
@@ -304,7 +234,9 @@ export async function flash(
 		analytics.logEvent('Write failed', eventData);
 		throw error;
 	}
+
 	windowProgress.clear();
+
 	if (flashState.wasLastFlashCancelled()) {
 		const eventData = {
 			...analyticsData,
@@ -327,6 +259,7 @@ export async function flash(
 
 /**
  * @summary Cancel write operation
+ * //TODO: find a better solution to handle cancellation
  */
 export async function cancel(type: string) {
 	const status = type.toLowerCase();
@@ -341,15 +274,7 @@ export async function cancel(type: string) {
 	};
 	analytics.logEvent('Cancel', analyticsData);
 
-	// Re-enable lock release on inactivity
-
-	try {
-		// @ts-ignore (no Server.sockets in @types/node-ipc)
-		const [socket] = ipc.server.sockets;
-		if (socket !== undefined) {
-			ipc.server.emit(socket, status);
-		}
-	} catch (error: any) {
-		analytics.logException(error);
+	if (cancelEmitter) {
+		cancelEmitter(status);
 	}
 }
