+++ lib/gui/app/pages/main/MainPage.tsx
@@ -26,10 +26,8 @@ import styled from 'styled-components';
 import FinishPage from '../../components/finish/finish';
 import { ReducedFlashingInfos } from '../../components/reduced-flashing-infos/reduced-flashing-infos';
 import { SettingsModal } from '../../components/settings/settings';
-import {
-	SourceMetadata,
-	SourceSelector,
-} from '../../components/source-selector/source-selector';
+import { SourceSelector } from '../../components/source-selector/source-selector';
+import { SourceMetadata } from '../../../../shared/typings/source-selector';
 import * as flashState from '../../models/flash-state';
 import * as selectionState from '../../models/selection-state';
 import * as settings from '../../models/settings';
