+++ lib/shared/catalina-sudo/sudo.ts
@@ -19,7 +19,7 @@ import { join } from 'path';
 import { env } from 'process';
 import { promisify } from 'util';
 
-import { getAppPath } from '../utils';
+import { getAppPath } from '../get-app-path';
 import { supportedLocales } from '../../gui/app/i18n';
 
 const execFileAsync = promisify(execFile);
