+++ lib/shared/utils.ts
@@ -14,9 +14,6 @@
  * limitations under the License.
  */
 
-import axios from 'axios';
-import { Dictionary } from 'lodash';
-
 import * as errors from './errors';
 
 export function isValidPercentage(percentage: any): boolean {
@@ -36,19 +33,6 @@ export async function delay(duration: number): Promise<void> {
 	await new Promise((resolve) => {
 		setTimeout(resolve, duration);
 	});
-}
-
-export function getAppPath(): string {
-	return (
-		(require('electron').app || require('@electron/remote').app)
-			.getAppPath()
-			// With macOS universal builds, getAppPath() returns the path to an app.asar file containing an index.js file which will
-			// include the app-x64 or app-arm64 folder depending on the arch.
-			// We don't care about the app.asar file, we want the actual folder.
-			.replace(/\.asar$/, () =>
-				process.platform === 'darwin' ? '-' + process.arch : '',
-			)
-	);
 }
 
 export function isJson(jsonString: string) {
