+++ lib/util/drive-scanner.ts
